# configuration-service/swagger.yaml
swagger: '2.0'
info:
  title: Configuration Service API
  version: 0.1.0
basePath: /v1
produces:
- application/json
consumes:
  - application/cloudevents+json
  - application/json
schemes:
  - http
definitions:
  Error:
    type: object
    required:
      - message
    properties: 
      code:
        type: integer
        format: int64
        description: Error code
      message:
        type: string
        description: Error message
...
  ExpandedProjects:
    type: object
    properties:
      nextPageKey:
        type: string
        description: Pointer to next page, base64 encoded
      totalCount:
        type: number
        description: Total number of projects
      pageSize:
        type: number
        description: Size of returned page
      projects:
        type: array
        items:
          $ref: '#/definitions/ExpandedProject'
  ExpandedProject:
    type: object
    properties:
      projectName:
        type: string
        description: Project name
      gitUser:
        type: string
        description: Git User
      gitRemoteURI:
        type: string
        description: Git remote URI
      shipyard:
        type: string
        description: Shipyard file content
      creationDate:
        type: string
        description: Creation date of the project
      lastEventContext:
        $ref: '#/definitions/EventContext'
      stages:
        type: array
        items:
          $ref: '#/definitions/ExpandedStage'
  ExpandedStage:
    type: object
    properties:
      stageName:
        type: string
        description: Stage name
      lastEventContext:
        $ref: '#/definitions/EventContext'
      services:
        type: array
        items:
          $ref: '#/definitions/ExpandedService'
  ExpandedService:
    type: object
    properties:
      serviceName:
        type: string
        description: Service name
      deployedImage:
        type: string
        description: Currently deployed image
      creationDate:
        type: string
        description: Creation date of the service
      lastEventTypes:
        type: object
        additionalProperties:
          $ref: '#/definitions/EventContext'
...
  EventContext:
    type: object
    properties:
      keptnContext:
        type: string
        description: Keptn Context ID of the event
      eventId:
        type: string
        description: ID of the event
      time:
        type: string
        description: Time of the event
  Project:
    type: object
    properties:
      projectName:
        type: string
        description: Project name
      gitUser:
        type: string
        description: Git User
      gitRemoteURI:
        type: string
        description: Git remote URI
      gitToken:
        type: string
        description: Git token
      stages:
        type: array
        items:
          $ref: '#/definitions/Stage'
...
  Stage:
    type: object
    properties:
      stageName:
        type: string
        description: Stage name
      services:
        type: array
        items:
          $ref: '#/definitions/Service'
  Stages:
    type: object
    properties:
      nextPageKey: 
        type: string
        description: Pointer to next page, base64 encoded
      totalCount: 
        type: number
        description: Total number of stages
      pageSize: 
        type: number
        description: Size of returned page
      stages:
        type: array
        items:
          $ref: '#/definitions/ExpandedStage'
  Service:
    type: object
    properties:
      serviceName:
        type: string
        description: Service name
...
parameters:
  project:
    in: body
    name: project
    description: Project entity
    schema:
      $ref: '#/definitions/Project'
  projectName:
    in: path
    name: projectName
    required: true
    type: string
    description: Name of the project
  stage:
    in: body
    name: stage
    description: Stage entity
    schema:
      $ref: '#/definitions/Stage'
  stageName:
    in: path
    name: stageName
    required: true
    type: string
    description: Name of the stage
...
  disableUpstreamSync:
    in: query
    name: disableUpstreamSync
    required: false
    default: false
    type: boolean
    description: Disable sync of upstream repo before reading content
...
  pageSize:
    in: query
    name: pageSize
    type: integer
    minimum: 1
    maximum: 50
    default: 20
    description: The number of items to return
  nextPageKey:
    in: query
    name: nextPageKey
    type: string
    description: Pointer to the next set of items
paths:
  '/project':
    post:
      tags:
        - Project
      deprecated: true
      summary: 'INTERNAL Endpoint: Create a new project by project name'
      parameters:
        - $ref: '#/parameters/project'
      responses:
        '204':
          description: Success. Project has been created. Response does not have a body.
        '400':
          description: Failed. Project could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Project
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of projects
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ExpandedProjects'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
  '/project/{projectName}':
    parameters:
      - $ref: '#/parameters/projectName'
    get:
      tags:
        - Project
      summary: Get the specified project
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ExpandedProject'
        '404':
          description: Failed. Project could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'  
    put:
      tags:
        - Project
      deprecated: true
      summary: 'INTERNAL Endpoint: Update the specified project'
      parameters:
        - $ref: '#/parameters/project'
      responses:
        '204':
          description: Success. Project has been updated. Response does not have a body.
        '400':
          description: Failed. Project could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    delete:
      tags:
        - Project
      deprecated: true
      summary: 'INTERNAL Endpoint: Delete the specified project'
      responses:
        '204':
          description: Success. Project has been deleted. Response does not have a body.
        '400':
          description: Failed. Project could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...
  '/project/{projectName}/stage':
    parameters:
      - $ref: '#/parameters/projectName'
    post:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Create a new stage by stage name'
      parameters:
        - $ref: '#/parameters/stage'
      responses:
        '204':
          description: Success. Stage has been created. Response does not have a body.
        '400':
          description: Failed. Stage could not be created.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    get:
      tags:
        - Stage
      parameters:
        - $ref: '#/parameters/pageSize'
        - $ref: '#/parameters/nextPageKey'
        - $ref: '#/parameters/disableUpstreamSync'
      summary: Get list of stages
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/Stages'
        '404':
          description: Failed. Containing project could not be found.
          schema:
            $ref: '#/definitions/Error' 
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
  '/project/{projectName}/stage/{stageName}':
    parameters:
      - $ref: '#/parameters/projectName'
      - $ref: '#/parameters/stageName'
    get:
      tags:
        - Stage
      summary: Get the specified stage
      parameters:
        - $ref: '#/parameters/disableUpstreamSync'
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/ExpandedStage'
        '404':
          description: Failed. Project resource could not be found.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error' 
    put:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Update the specified stage'
      parameters:
        - $ref: '#/parameters/stage'
      responses:
        '204':
          description: Success. Stage has been updated. Response does not have a body.
        '400':
          description: Failed. Stage could not be updated.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
    delete:
      tags:
        - Stage
      deprecated: true
      summary: 'INTERNAL Endpoint: Delete the specified stage'
      responses:
        '204':
          description: Success. Stage has been deleted. Response does not have a body.
        '400':
          description: Failed. Stage could not be deleted.
          schema:
            $ref: '#/definitions/Error'
        'default':
          description: Error
          schema:
            $ref: '#/definitions/Error'
...